projetPFA-app/projetPFA/requirements.txt
streamlit==1.28.0
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
plotly==5.15.0
Pillow==10.0.0